fix: restore semantics lost during deduplication
Code review found 3 regressions + 1 CSS parity drift:

1. VariantCardGrid: restore role=button, tabindex=0, and keyboard
   handlers (Enter/Space) that existed in ModalDialog's inline cards

2. useModalOverlay: add event.stopImmediatePropagation() on Escape
   to prevent bubbling to other document-level listeners when
   multiple modals are stacked

3. standaloneMigration.restoreSnapshotIntoInstallation: add missing
   restoreComfyUIVersion + buildPostRestoreState steps, and use
   updated inst state when saving post-restore snapshot. Also
   replace 40-line inline duplicate in registerInstallationHandlers
   with a call to the now-complete shared function.

4. SnapshotTab: scoped override for ls-node-disabled to preserve
   original --text-faint color (global ls- class uses --text-muted)

Amp-Thread-ID: https://ampcode.com/threads/T-019db7b6-3400-7293-a487-351dcca1977b
