+++ src/main/lib/ipc/registerInstallationHandlers.ts
@@ -6,13 +6,12 @@ import {
   syncOemSeedBestEffort, isEffectivelyEmptyInstallDir,
   download, createCache, extract, deleteDir, formatDeleteStatus, deleteAction, untrackAction,
   MARKER_FILE,
-  validateExportEnvelope, importSnapshots, saveSnapshot, getSnapshotCount,
-  restoreCustomNodes, restorePipPackages, restoreComfyUIVersion, buildPostRestoreState,
   _operationAborts,
   sanitizeEnvVars,
   getComfyArgsSchema,
 } from './shared'
 import type { ComfyVersion, ComfyArgDef } from './shared'
+import { restoreSnapshotIntoInstallation } from '../standaloneMigration'
 
 export function registerInstallationHandlers(): void {
   // Installations
@@ -175,50 +174,11 @@ export function registerInstallationHandlers(): void {
           }
           const update = (data: Record<string, unknown>): Promise<void> =>
             installations.update(installationId, data).then(() => {})
-
-          try {
-            const fileContent = await fs.promises.readFile(pendingFile, 'utf-8')
-            const envelope = validateExportEnvelope(JSON.parse(fileContent))
-            await importSnapshots(freshInst.installPath, envelope)
-            const targetSnapshot = envelope.snapshots[0]!
-
-            // Restore ComfyUI version
-            sendOutput('\n── Restore ComfyUI Version ──\n')
-            const comfyResult = await restoreComfyUIVersion(freshInst.installPath, targetSnapshot, sendOutput)
-
-            sendOutput('\n── Restore Nodes ──\n')
-            await restoreCustomNodes(freshInst.installPath, freshInst, targetSnapshot, sendProgress, sendOutput, abort.signal, settings.getMirrorConfig())
-
-            if (!abort.signal.aborted && !targetSnapshot.skipPipSync) {
-              sendOutput('\n── Restore Packages ──\n')
-              await restorePipPackages(freshInst.installPath, freshInst, targetSnapshot,
-                (phase, data) => sendProgress(phase === 'restore' ? 'restore-pip' : phase, data),
-                sendOutput, abort.signal, settings.getMirrorConfig())
-            }
-
-            const restoreState = buildPostRestoreState(
-              targetSnapshot, comfyResult,
-              freshInst.updateInfoByChannel as Record<string, Record<string, unknown>> | undefined,
-              freshInst.comfyVersion as ComfyVersion | undefined
-            )
-            await update(restoreState)
-
-            // Save post-restore snapshot
-            try {
-              const updatedInst = { ...freshInst, ...restoreState }
-              const filename = await saveSnapshot(freshInst.installPath, updatedInst, 'post-restore')
-              const snapshotCount = await getSnapshotCount(freshInst.installPath)
-              await update({ pendingSnapshotRestore: undefined, lastSnapshot: filename, snapshotCount })
-            } catch {
-              await update({ pendingSnapshotRestore: undefined })
-            }
-          } catch (restoreErr) {
-            console.warn('Post-install snapshot restore failed:', restoreErr)
-            sendOutput(`\n⚠ Snapshot restore failed: ${(restoreErr as Error).message}\nThe installation completed successfully. You can restore the snapshot manually from the Snapshots tab.\n`)
-            await update({ pendingSnapshotRestore: undefined })
-          } finally {
-            fs.promises.unlink(pendingFile).catch(() => {})
-          }
+          await restoreSnapshotIntoInstallation(
+            freshInst, pendingFile, true,
+            { sendProgress, sendOutput, signal: abort.signal },
+            update,
+          )
         }
 
         sendProgress('done', { percent: 100, status: 'Complete' })
