+++ src/renderer/src/components/VariantCardGrid.vue
@@ -17,11 +17,15 @@ defineEmits<{
     <div
       v-for="opt in options"
       :key="opt.value"
+      role="button"
+      tabindex="0"
       :class="['variant-card', {
         selected: selectedValue === opt.value,
         recommended: opt.recommended,
       }]"
       @click="$emit('select', opt)"
+      @keydown.enter.prevent="$emit('select', opt)"
+      @keydown.space.prevent="$emit('select', opt)"
     >
       <div class="variant-card-icon">
         <img
