+++ src/renderer/src/composables/useModalOverlay.ts
@@ -27,6 +27,7 @@ export function useModalOverlay(
 
   function handleEscapeKey(event: KeyboardEvent): void {
     if (event.key === 'Escape' && shouldClose()) {
+      event.stopImmediatePropagation()
       close()
     }
   }
